test: split API ordering integration tests
