+++ backend/tests/api_ordering_test_setup.js
@@ -1,0 +1,29 @@
+const expressApp = require("../src/express_app");
+const { addRoutes } = require("../src/server");
+const { getMockedRootCapabilities } = require("./spies");
+const {
+    stubEnvironment,
+    stubLogger,
+    stubDatetime,
+    stubEventLogRepository,
+} = require("./stubs");
+
+/**
+ * Creates an Express app configured with mocked capabilities for testing.
+ * @returns {Promise<{app: import('express').Express, capabilities: any}>}
+ */
+async function makeTestApp() {
+    const capabilities = getMockedRootCapabilities();
+    stubEnvironment(capabilities);
+    stubLogger(capabilities);
+    stubDatetime(capabilities);
+    await stubEventLogRepository(capabilities);
+
+    const app = expressApp.make();
+    capabilities.logger.enableHttpCallsLogging(app);
+    await addRoutes(capabilities, app);
+
+    return { app, capabilities };
+}
+
+module.exports = { makeTestApp };
